Refactor to use slices instead of passing Vecs
Thanks to @etam@im-in.space on Mastodon for the tip on how to convert
multi-dimension nested Vecs to slices!

https://im-in.space/@etam/113204943591106746

Code was adapted so that the final sliced result is contained within a
Vec, so the result becomes Vec<&[&[u32]]>. Because if we try and pass a
&[&[&[u32]]] to the function, the variable doesn't live long enough and
the compiler complains. So store the values in a Vec, and pass by
reference.